feat: Respond to user input
