--- src/main/java/Bee.java
@@ -1,16 +1,0 @@
-public class Bee {
-    public static void main(String[] args) {
-        String logo = " ____                 \n"
-                + "|  _ \\  ___   ___ \n"
-                + "| |_/  / _ \\ / _ \\\n"
-                + "| |_\\ |  __/|  __/ \n"
-                + "|____/ \\___| \\___|\n";
-        System.out.println(logo);
-        System.out.println("____________________________________________________________");
-        System.out.println(" Hello, I'm Bee!");
-        System.out.println(" What can I do for you?\n");
-        System.out.println("____________________________________________________________");
-        System.out.println(" Bye. Hope to see you again soon!");
-        System.out.println("____________________________________________________________");
-    }
-}
